Fix for “More than one plugin attempted to publish the same registration name `onError` when using RN 0.39
This also fixes the event handlers for Android. The handlers defined `YouTube.android.js` were never triggered because the event names in the .java file were different. By accident the handlers passed to the YouTube component were still called, but directly by the native-code. This caused the arguments to be wrong. This commit fixes that as well.