Use a more reliable way of getting the currentActivity - it used to crash occasionally
When the PlayerWindowActivity is minimised, the Navigation component was calling 'getReactInstanceManager().onHostResume(activity' which tells the React Native instance manager that the screen below the player is now the current activity which technically it is. However, the YouTube player was asking the React Native instance manager to tell it the currentActivity so it know who was managing it's fragment. This change now tells the YouTube player that it's root view is it's current view. In practice this is PlayerWindowActivity.